+++ packages/amplify-graphql-transformer-core/src/config/index.ts
@@ -1,0 +1,11 @@
+export { TransformerProjectConfig } from './project-config';
+export {
+  TransformConfig,
+  ConflictDetectionType,
+  ConflictHandlerType,
+  ResolverConfig,
+  SyncConfig,
+  SyncConfigLambda ,
+  SyncConfigOptimistic,
+  SyncConfigServer,
+} from './transformer-config';