+++ packages/amplify-graphql-transformer-core/src/transformation/index.ts
@@ -1,0 +1,2 @@
+export { GraphQLTransform, GraphQLTransformOptions } from './transform'
+export *  from './transformer-config';