+++ packages/graphql-relational-schema-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.19](https://github.com/aws-amplify/amplify-cli/compare/graphql-relational-schema-transformer@2.15.18...graphql-relational-schema-transformer@2.15.19) (2020-11-08)
+
+**Note:** Version bump only for package graphql-relational-schema-transformer
+
+
+
+
+
 ## [2.15.18](https://github.com/aws-amplify/amplify-cli/compare/graphql-relational-schema-transformer@2.15.17...graphql-relational-schema-transformer@2.15.18) (2020-10-30)